+++ fight_logs.js
@@ -1,94 +1,55 @@
 // Подлючаем функции
-import {getTime} from "./utils.js";
-import getRandom from "./utils.js";
+// import {getTime} from './utils/index.js';
+// import getRandom from './utils/index.js';
+// import { LOGS } from './constants/index.js';
 
-const $chat = document.querySelector('.chat');
-
-const logs = {
-    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-    end: [
-        'Результат удара [playerWins]: [playerLose] - труп',
-        '[playerLose] погиб от удара бойца [playerWins]',
-        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-    ],
-    hit: [
-        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-    ],
-    defence: [
-        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-    ],
-    draw: 'Ничья - это тоже победа!'
-};
+// const 
 
 // Деструктурируем объект logs. Получаем сообщение для лога
-const {start, end, hit, defence, draw} = logs;
-export function getLogMessage(type, player1, player2) {
+// export function getLogMessage(type, player1, player2) {
+//     const {start, end, hit, defence, draw} = LOGS;
     
-    let text;
-    switch(type) {
-        case 'start' :
-            text = start
-                .replace('[player1]', player1)
-                .replace('[player2]', player2);
-            return text;
-        case 'hit' :
-            text = hit[getRandom(hit.length) - 1]
-                .replace('[playerKick]', player1)
-                .replace('[playerDefence]', player2);
-            return text;
-        case 'defence' :
-            text = defence[getRandom(defence.length) - 1]
-                .replace('[playerKick]', player1)
-                .replace('[playerDefence]', player2);
-            return text;
-        case 'end' :
-            text = end[getRandom(end.length) - 1]
-                .replace('[playerWins]', player1)
-                .replace('[playerLose]', player2);
-            return text;
-        case 'draw' :
-            text = draw;
-            return text;
-        default :
-            text = '... ждем ...';
-            return text;
-    }
-};
+//     let text;
+//     switch(type) {
+//         case 'start' :
+//             text = start
+//                 .replace('[player1]', player1)
+//                 .replace('[player2]', player2);
+//             return text;
+//         case 'hit' :
+//             text = hit[getRandom(hit.length) - 1]
+//                 .replace('[playerKick]', player1)
+//                 .replace('[playerDefence]', player2);
+//             return text;
+//         case 'defence' :
+//             text = defence[getRandom(defence.length) - 1]
+//                 .replace('[playerKick]', player1)
+//                 .replace('[playerDefence]', player2);
+//             return text;
+//         case 'end' :
+//             text = end[getRandom(end.length) - 1]
+//                 .replace('[playerWins]', player1)
+//                 .replace('[playerLose]', player2);
+//             return text;
+//         case 'draw' :
+//             text = draw;
+//             return text;
+//         default :
+//             text = '... ждем ...';
+//             return text;
+//     }
+// };
 
 // Генерируем информацию для лога
-export function generateLogs(text, damage, playerHP) {
-    const fightTime = getTime();
-    const message = text.replace('[time]', fightTime);
-    let el;
-    if (damage) {
-        el = `<p>${fightTime} ${message} -${damage} ${playerHP}/100</p>`;
-    } else {
-        el = `<p>${fightTime} ${message}</p>`;
-    }
+// export function generateLogs(text, damage, playerHP) {
+//     const fightTime = getTime();
+//     const message = text.replace('[time]', fightTime);
+//     let el;
+//     if (damage) {
+//         el = `<p>${fightTime} ${message} -${damage} ${playerHP}/100</p>`;
+//     } else {
+//         el = `<p>${fightTime} ${message}</p>`;
+//     }
 
-    $chat.insertAdjacentHTML('afterbegin', el);
-};
+//     $chat.insertAdjacentHTML('afterbegin', el);
+// };
